Evaluate a simulation function on each row of a data frame

Evaluates a simulation function on each row of a data frame containing parameter values. Returns a single data frame with parameters and simulation results. The function uses furrr::future_pmap, which allows for easy parallelization.

evaluate_by_row(
  params,
  sim_function,
  ...,
  .progress = FALSE,
  .options = furrr::future_options(),
  system_time = TRUE
)

Arguments

params
data frame or tibble containing simulation parameter values. Each row should represent a separate set of parameter values.

sim_function
function to be evaluated, with argument names matching the variable names in params. The function must return a data.frame, tibble, or vector.

...
additional arguments passed to sim_function.

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans.

.options
The future specific options to use with the workers. This must be the result from a call to future_options().

system_time
logical indicating whether to print computation time. TRUE by default.

Value

A tibble containing parameter values and simulation results.

Examples

df <- data.frame(
  n = 3:5,
  lambda = seq(8, 16, 4)
)

evaluate_by_row(df, rpois)

#>    user  system elapsed 
#>   0.077   0.004   0.081 

#> # A tibble: 12 x 3
#>        n lambda .results
#>    <int>  <dbl>    <int>
#>  1     3      8        4
#>  2     3      8       11
#>  3     3      8        9
#>  4     4     12        8
#>  5     4     12       11
#>  6     4     12       14
#>  7     4     12       15
#>  8     5     16        8
#>  9     5     16       12
#> 10     5     16       14
#> 11     5     16       13
#> 12     5     16       12