Repeat an expression multiple times and (optionally) stack the results.

Repeat an expression (usually involving random number generation) multiple times. Optionally, organize the results into a data.frame that stacks the output from all replications of the expression.

Usage

repeat_and_stack(n, expr, stack = TRUE)

Arguments

n

Number of times to repeat the expression

expr

An expression to be evaluated.

stack

Logical value indicating whether to organize the results into a data.frame.

Value

If stack = TRUE (the default), the results of each evaluation of expr will be stacked together using rbind. If stack = FALSE, a list of length n with entries corresponding to the output of each replication of expr.

Examples

repeat_and_stack(n = 3, data.frame(x = rexp(2)))
#>           x
#> 1 1.1867656
#> 2 2.0933169
#> 3 0.7543366
#> 4 1.4212582
#> 5 0.3884923
#> 6 1.7279031

repeat_and_stack(n = 3, data.frame(x = rexp(2)), stack = FALSE)
#> [[1]]
#>           x
#> 1 0.4464174
#> 2 0.7105519
#> 
#> [[2]]
#>            x
#> 1 0.09738847
#> 2 1.39750947
#> 
#> [[3]]
#>           x
#> 1 0.6713126
#> 2 0.8858652
#>